+++ lib/src/extensions/helpers_extension.dart
@@ -195,6 +195,17 @@ extension StringUtility2 on String? {
     return DateTime.now().subtract(time);
   }
 
+  Uri? toUri() {
+    if (this == null) {
+      return null;
+    }
+    try {
+      return Uri.parse(this!);
+    } on FormatException {
+      return null;
+    }
+  }
+
   DateTime? tryParseDateTime() {
     if (this == null) {
       return null;
@@ -291,7 +302,7 @@ extension GetOrNullMap on Map {
         return null;
       }
     }
-    return value as T;
+    return value as T?;
   }
 }